+++ lib/feature/displaying_graphs/boundary/usecase/charts_usecase.dart
@@ -1,4 +1,4 @@
-import 'package:ics/domain/entity/point.dart';
+import '../../domain/entity/point.dart';
 
 abstract class ChartsUseCase {
   Future<List<List<Point>>> buildTriangle(List<double> data, int countTerm);
