+++ lib/feature/displaying_graphs/domain/entity/plenty.dart
