+++ lib/feature/displaying_graphs/domain/entity/point.dart
