+++ lib/feature/uploading_data/domain/entity/chart_params.dart
@@ -1,0 +1,6 @@
+class ChartParams {
+  final int chartIndex;
+  final int termCount;
+
+  ChartParams({required this.chartIndex, required this.termCount});
+}
